sui_edge_proxy::config
Struct ProxyConfig

pub struct ProxyConfig {
    pub listen_address: SocketAddr,
    pub metrics_address: SocketAddr,
    pub execution_peer: PeerConfig,
    pub read_peer: PeerConfig,
    pub max_idle_connections: usize,
    pub idle_timeout_seconds: Duration,
    pub logging: LoggingConfig,
}

Fields
listen_address: SocketAddrmetrics_address: SocketAddrexecution_peer: PeerConfigread_peer: PeerConfigmax_idle_connections: usize

Maximum number of idle connections to keep in the connection pool. When set, this limits the number of connections that remain open but unused, helping to conserve system resources.

idle_timeout_seconds: Duration

Idle timeout for connections in the connection pool. This should be set to a value less than the keep-alive timeout of the server to avoid sending requests to a closed connection. if your you expect sui-edge-proxy to recieve a small number of requests per second, you should set this to a higher value.

logging: LoggingConfig

Logging configuration for read requests including sample rate and log file path.